UI: extract the Windows specific bits of Color
Make Color more generic by externalising the Windows specific handling
of color where possible.
